[Refactor] :rainbow: notify todosSubject when new todoItem added #9
기존의 todolist에 바로 html을 추가하는 것이 아닌, 옵저버 패턴을 이용해
서브젝트(컨트롤러) 쪽으로 데이터 추가 및 수정 로직을 옮김
